front/requirements.txt
streamlit
plotly
mysql-connector-python
cryptography
boto3
pyupbit